+++ src/Model/Produto.java
@@ -11,7 +11,7 @@ package Model;
  */
 public class Produto {
     private String nome;
-    private double preco;
+    private float preco;
     private String tipoproduto;
     //testando
     public String getNome() {
@@ -22,11 +22,11 @@ public class Produto {
         this.nome = nome;
     }
 
-    public double getPreco() {
+    public float getPreco() {
         return preco;
     }
 
-    public void setPreco(double preco) {
+    public void setPreco(float preco) {
         this.preco = preco;
     }
